requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
brotli=1.0.9=ha925a31_2
ca-certificates=2022.4.26=haa95532_0
certifi=2022.6.15=py39haa95532_0
cffi=1.15.0=py39h2bbff1b_1
colorama=0.4.4=pyhd3eb1b0_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
future=0.18.2=py39haa95532_1
icu=58.2=ha925a31_3
intel-openmp=2021.4.0=haa95532_3556
jpeg=9e=h2bbff1b_0
kiwisolver=1.4.2=py39hd77b12b_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=he0120a3_1
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
matplotlib=3.5.1=py39haa95532_1
matplotlib-base=3.5.1=py39hd77b12b_1
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
munkres=1.1.4=py_0
ninja=1.10.2=haa95532_5
ninja-base=1.10.2=h6d14046_5
numpy=1.22.3=py39h7a0a035_0
numpy-base=1.22.3=py39hca35cd5_0
openssl=1.1.1o=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pillow=9.0.1=py39hdc2b20a_0
pip=21.2.4=py39haa95532_0
pycparser=2.21=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39hd77b12b_6
python=3.9.12=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.10.2=cpu_py39h907fbb5_0
qt=5.9.7=vc14h73c81de_0
setuptools=61.2.0=py39haa95532_0
sip=4.19.13=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.5=h2bbff1b_0
tk=8.6.12=h2bbff1b_0
torchvision=0.11.3=py39_cu113
tornado=6.1=py39h2bbff1b_0
tqdm=4.64.0=py39haa95532_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py39haa95532_2
xz=5.2.5=h8cc25b3_1
zlib=1.2.12=h8cc25b3_2
zstd=1.5.2=h19a0ad4_0
pytorch-fid
lpips
image-quality

yaml~=0.2.5
pyyaml~=6.0
torchvision~=0.11.3
einops~=0.4.1
tqdm~=4.64.0